Add comparison tool test runner, built on mininode
comptool.py creates a tool for running a test suite on top of the
mininode p2p framework.  It supports two types of tests: those for
which we expect certain behavior (acceptance or rejection of a block
or transaction) and those for which we are just comparing that the
behavior of 2 or more nodes is the same.

blockstore.py defines BlockStore and TxStore, which provide leveldb-
backed maps between block/tx hashes and the corresponding block or
tx.

blocktools.py defines utility functions for creating and manipulating
blocks and transactions.

invalidblockrequest.py is an example test in the comptool framework,
which tests the behavior of a single node when sent two different
types of invalid blocks (a block with a duplicated transaction and
a block with a bad coinbase value).