Add script manipulation tools for use in mininode testing framework
script.py is modified from the code in python-bitcoinlib, and provides
tools for manipulating and creating CScript objects.

bignum.py is a dependency for script.py (which could potentially be
removed)

script_test2.py is an example script that uses the script tools for
running a test that compares the behavior of two nodes (in a comptool
test) for each of the test cases in the bitcoin unit test script
files (script_valid.json and script_invalid.json).