Add script manipulation tools for use in mininode testing framework
script.py is modified from the code in python-bitcoinlib, and provides
tools for manipulating and creating CScript objects.

bignum.py is a dependency for script.py

script_test.py is an example test that uses the script tools for
running a test that compares the behavior of two nodes, in a comptool-
style test, for each of the test cases in the bitcoin unit test script
files, script_valid.json and script_invalid.json.  (This test is
very slow to run, but is a proof of concept for how we can write tests
to compare consensus-critical behavior between different versions of
bitcoind.)

bipdersig-p2p.py is another example test in the comptool framework,
which tests deployment of BIP DERSIG for a single node.  It uses
the script.py tools for manipulating signatures to be non-DER
compliant.