Rewrite CreateNewBlock
Use the score index on the mempool to only add sorted txs in order.  Remove much of the validation while building the block, relying on mempool to be consistent and only contain txs that can be mined.
The mempool is assumed to be consistent as far as not containing txs which spend non-existent outputs or double spends, and scripts are valid.  Finality of txs is still checked (except not coinbase maturity, assumed in mempool).
Still TestBlockValidity in case mempool consistency breaks and change to log error and return NULL if an invalid block was created.
Unit tests are modified to realize that invalid blocks can now be constructed if the mempool breaks its consitency assumptions.

Conflicts:
	src/miner.cpp
(hidden) src/test/miner_tests.cpp