Rework mempool handling by datalogger and simulation
Now that upstream supports writing a mempool on shutdown, we can eliminate
recording the mempool by the datalogger.  Also scrap our code for loading
the mempool (which was broken, being reliant on a writer which wouldn't write
transactions in dependency order) in favor of the LoadMempool() function now
available in main.cpp

Fix one small bug in LoadMempool(), where free-tx rate limiting was kicking in
for in-mempool transactions.