[qa] Handle potential cookie race when starting node
When a node is restarted during a test after an unclean shutdown (such
as with -dbcrashratio), it's possible an old cookie file was left
behind. This can cause a race condition when restarting the node, where
the test framework might try to connect using credentials from the
old cookie file, just as the node will generate new credentials and
overwrite the old file.

If this happens, the test framework will receive an 'Unauthorized'
response from the server. Handle this the same way we deal with missing
credentials (sleep and try again).