Fix concurrency-related bugs in ActivateBestChain
If multiple threads are invoking ActivateBestChain, it was possible to have
them working towards different tips, and we could arrive at a less work tip
than we should.  Fix this by updating pindexMostWork if our tip has changed
since the last loop iteration.  As belt and suspenders, we also check that
pindexMostWork (where we're working towards) must still be in
setBlockIndexCandidates, or else we recalculate it.

Also, previously we could release cs_main at a time when our tip has less work
than a previous point if we ever tried to connect an invalid block.  Fix this
by holding on to cs_main and calling ActivateBestChainStep again if an invalid
block is found.