Add chain-sync peer connections
This introduces a new category of outbound peer connections for syncing
our chain tip.

Once our tip is synced or nearly synced, periodically make connections to other
peers for the purpose of exchanging headers (and, if we learn of a new block,
downloading such blocks). Only stay connected for a short period of time, and
then disconnect.

Do not relay transactions with such peers, as these should be low-bandwidth,
short-lived connections. These are designed to make eclipse attacks more
difficult to pull off, without incurring a very high cost in bandwidth or use
of network resources.