Fix validation_chainstate_tests
This fixes the validation_chainstate_tests, which were left broken in the prior
commit.

Some of the invariants in CheckBlockIndex were modified, but more work is
needed (ie to move CheckBlockIndex to ChainstateManager, as most of what
CheckBlockIndex is doing is checking the consistency of the block index, which
is outside of Chainstate).