ultrahack for a tx replacing another tx with same parents
In the case that a single transaction is replacing another single transaction
with the exact same parents, seed the initial sort order for the replacement
cluster with the original cluster, substituting the new transaction in for the
old.

The hope is that we can avoid pathological situations where the linearization
algorithm treats a strictly better replacement as being worse, due to quirks in
ancestor feerate sorting.