Benchmark linearization algorithm
Every we time we relinearize a cluster, run the linearization algorithm 10
times with no pre-existing linearization, and track total system time used.

If the total system time exceeds the highest time observed so far for a cluster
with the same number of transactions, output it.

Do the same with 10 iterations of the linearization algorithm using an optimal
linearization as input.